1.添加ImGuizmo，但目前无法显示控件 2.修复部分警告

+++ CraftNow/src/CraftNow/Renderer/Texture.cpp
@@ -38,7 +38,7 @@ namespace CraftNow {
 		uint8_t* data = nullptr;
 		uint64_t size = 0;
 		//stbi_set_flip_vertically_on_load(1);
-		data = stbi_load_from_memory((const stbi_uc*)buffer, length, &width, &height, &channels, 4);
+		data = stbi_load_from_memory((const stbi_uc*)buffer, static_cast<int>(length), &width, &height, &channels, 4);
 		size = width * height * 4;
 
 		outWidth = width;
