+++ CraftNow/src/CraftNow/Utils/ChineseUtils.cpp
@@ -8,7 +8,7 @@ namespace CraftNow {
 		int nwLen = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), -1, NULL, 0);
 		wchar_t* pwBuf = new wchar_t[nwLen + 1];
 		memset(pwBuf, 0, nwLen * 2 + 2);
-		MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), pwBuf, nwLen);
+		MultiByteToWideChar(CP_UTF8, 0, str.c_str(), static_cast<int>(str.length()), pwBuf, nwLen);
 
 		int nLen = WideCharToMultiByte(CP_ACP, 0, pwBuf, -1, NULL, NULL, NULL, NULL);
 		char* pBuf = new char[nLen + 1];
@@ -27,7 +27,7 @@ namespace CraftNow {
 		int nwLen = MultiByteToWideChar(CP_ACP, 0, str.c_str(), -1, NULL, 0);
 		wchar_t* pwBuf = new wchar_t[nwLen + 1];
 		memset(pwBuf, 0, nwLen * 2 + 2);
-		MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), pwBuf, nwLen);
+		MultiByteToWideChar(CP_ACP, 0, str.c_str(), static_cast<int>(str.length()), pwBuf, nwLen);
 
 		int nLen = WideCharToMultiByte(CP_UTF8, 0, pwBuf, -1, NULL, NULL, NULL, NULL);
 		char* pBuf = new char[nLen + 1];
